requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.0
requests==2.32.4
schedule==1.2.2
urllib3==2.4.0
